requirements.txt
acdh-baserow-client
coverage
fastapi[all]
fastapi-cache2
flake8
pytest